Guard against Flymake errors with invalid buffer positions.
If a backend function provides an invalid buffer position to
`flymake-diag-region' that function will return nil. When Flymake
processes the diagnostic it tries to create an overlay with this
position, signaling an error to the user. We guard against this by
filtering out such diagnostics but recording the error in the Flymake
log.